tests/wpt/web-platform-tests/bluetooth/service/getCharacteristics/gen-service-is-removed.https.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Service is removed before getCharacteristics call. ' +
    'Reject with InvalidStateError.';
const expected = new DOMException('GATT Service no longer exists.',
    'InvalidStateError');
let service, fake_service, fake_peripheral;

bluetooth_test(() => getHealthThermometerService()
    .then(_ => ({service, fake_service, fake_peripheral} = _))
    .then(() => fake_service.remove())
    .then(() => fake_peripheral.simulateGATTServicesChanged())
    .then(() => assert_promise_rejects_with_message(
        service.getCharacteristics(),
        expected,
        'Service got removed.')),
    test_desc);

</script>
